check stock prices

+++ app/models/alert.rb
@@ -11,7 +11,7 @@ class Alert < ApplicationRecord
     sp_index_open = data[:sp_index_data].open
     Alert.all.each do |alert|
       if alert[:item] == 'stock'
-        price = self.get_stock_price alert[:stock]
+        price = self.get_stock_price alert[:stock].split(',')[0]
       elsif alert[:item] == 'sp_index'
         price = sp_index_open
       elsif alert[:item] == 'gold'
@@ -19,8 +19,7 @@ class Alert < ApplicationRecord
       elsif alert[:item] == 'silver'
         price = silver_price
       end
-
-      if (alert[:above_or_below] && (price > alert[:target_value])) || (!alert[:above_or_below] && (price < alert[:target_value]))
+      if (!price.blank?) && (alert[:above_or_below] && (price > alert[:target_value])) || (!alert[:above_or_below] && (price < alert[:target_value]))
         PriceAlertMailer.send_alert_email(alert, price).deliver
         alert.destroy!
       end
@@ -28,7 +27,7 @@ class Alert < ApplicationRecord
   end
 
   def self.get_stock_price(stock)
-    Quandl::Dataset.get('GOOG/' + stock).data(params: { limit: 1 }).first.open
+    Quandl::Dataset.get(stock).data(params: { limit: 1 }).first.open
   end
 
   def self.get_gold_silver_stock_data
